Fold moving_sphere.h into sphere.h
This is in preparation of adding the moving capability to the standard
sphere class.
