+++ CMakeLists.txt
@@ -45,7 +45,6 @@ set ( SOURCE_NEXT_WEEK
   src/TheNextWeek/hittable.h
   src/TheNextWeek/hittable_list.h
   src/TheNextWeek/material.h
-  src/TheNextWeek/moving_sphere.h
   src/TheNextWeek/quad.h
   src/TheNextWeek/scene.h
   src/TheNextWeek/sphere.h
