+++ src/TheNextWeek/main.cc
@@ -17,7 +17,6 @@
 #include "constant_medium.h"
 #include "hittable_list.h"
 #include "material.h"
-#include "moving_sphere.h"
 #include "quad.h"
 #include "scene.h"
 #include "sphere.h"
